requirements.txt
pandas==2.3.1
numpy==2.3.1
scikit-learn==1.7.1
oracledb==2.5.1
openpyxl==3.1.5
requests==2.32.4
beautifulsoup4==4.13.4
selenium==4.33.0
lxml==5.2.2
tqdm==4.66.4
python-dotenv==1.0.1
xgboost==2.0.3
seaborn==0.13.2
matplotlib==3.10.3

Flask==3.0.3
Flask-Migrate==4.0.7
SQLAlchemy==2.0.21
cx__oracle==8.3.0
itsdangerous==2.2.0
Werkzeug==3.0.3
Flask-WTF==1.2.2
email_validator==1.3.1
dnspython==2.6.1
idna==3.7
alembic==1.11.1
Flask-Cors==4.0.1
Flask-Login==0.6.3

llama-cpp-python
transformers
torch
tensorflow
huggingface_hub
langchain-core
openvino_genai
optimum.intel
tf-keras
nncf

chromadb
sentence-transformers
pypdf
requests